Arduino Code
Code for Arduino lock box. Using buttons to recreate C, D, E, and F notes, you would enter a secret combination that would activate the servo, which would reveal the padlock code and so the user can unlock the box. The sound was made from a speaker, and the circuit was run by an Arduino.